+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ### Changes
 - Allow to disable handling of remote signals ([#2](https://github.com/neptune-ai/neptune-client-experimental/pull/2))
 - Added support for multiple features ([#3](https://github.com/neptune-ai/neptune-client-experimental/pull/3))
+- Sample logging for series errors ([#4](https://github.com/neptune-ai/neptune-client-experimental/pull/4))